refactor: 리팩토링 유틸리티 함수 파일 분리
