+++ modules/utils/validationUtils.js
@@ -1,16 +1,4 @@
-import { Random } from '@woowacourse/mission-utils';
-import { RANGE_NUMBER } from './constant';
-
-const createRandomNumber = () => {
-  const randomNumber = Random.pickNumberInRange(
-    RANGE_NUMBER.randomNumberMax,
-    RANGE_NUMBER.randomNumberMax
-  );
-
-  return randomNumber;
-};
-
-const checkCanMove = () => createRandomNumber() >= RANGE_NUMBER.canMove;
+import { RANGE_NUMBER } from '../constant';
 
 const checkCarNamesHaveBlank = (carNames) =>
   carNames.includes('') || carNames.includes(' ');
@@ -43,8 +31,6 @@ const checkTryNumberIsCharacter = (userInput) =>
   Number.isNaN(Number(userInput)) || Number(userInput) === 0;
 
 export {
-  createRandomNumber,
-  checkCanMove,
   checkCarNamesHaveBlank,
   checkCarNameIsOverFive,
   checkCarNameWithBlank,
